chore: Move FutureExt out of trace into context to facilitate re-use
