+++ opentelemetry/src/trace/mod.rs
@@ -177,9 +177,7 @@ mod tracer;
 mod tracer_provider;
 
 pub use self::{
-    context::{
-        get_active_span, mark_span_as_active, FutureExt, SpanRef, TraceContextExt, WithContext,
-    },
+    context::{get_active_span, mark_span_as_active, FutureExt, SpanRef, TraceContextExt},
     span::{Span, SpanKind, Status},
     span_context::{SpanContext, TraceState},
     tracer::{SamplingDecision, SamplingResult, SpanBuilder, Tracer},
